Added MarketPress to 'coursepress/pro' branch
MarketPress added as a submodule to this branch, so there are a few
things to be aware of when switching branches.

When switching to another branch, including master, be sure to run:

    git clean -dff

The -d will clean directories, but to remove submodule repositories
you need to specify the -f flag twice.

When switching from another branch, run the same command as above,
but be sure to update the submodule to get them back:

    git submodule update

You might need to run it recursively, in which case you need to run:

    git submodule update --recursive