Removed directory scanning from TCPDF
tcpd.php contains a method getFontsList() that uses directory
scanning to retrieve the list of fonts.  Ideally, it would be better
to override this method, but its a protected method.

Method altered to use an array of predefined fonts instead.

Its important to keep a note of this if we ever update the external
library.