Fixed issue with thumbnail generation
The thumbnail code could be considered redundant because of the list
image. However, there are several shortcodes that allow for a
thumbnail to be displayed.

Rather than removing the shortcodes, the thumbnail code has been
fixed to generate a thumbnail from the featured image URL, as long
as it is a file thats been added to the library and not a URL to
another site.