Added course completion meta.
New actions added to Course_Completion[]

    `
    // Fires everytime a course is deemed to be completed (can be a number of times in a single page load)
    do_action( 'coursepress_set_course_completed', $this->student_id, $this->id );
    `

    `
    // Fires everytime a unit is deemed to be completed (can be a number of times in a single page load)
    do_action( 'coursepress_set_unit_completed', $this->student_id, $this->id, $unit->ID );
    `

New hooks added to Student[]. These make use of the above mentioned
actions to add user meta containing completion information. Checks to
make sure that meta is only written if a course/unit has not yet been
marked as complete.

    `
    add_course_completed_meta()
    `

    `
    add_unit_completed_meta()
    `

New actions added to Student[]. The following actions are called in
the two methods above (respectively). They are triggered the first
time a course or unit is marked as completed.

    `
    // Will only fire once when a course is marked as complete, should not trigger again.
    do_action( 'coursepress_student_course_completed', $student_id, $course_id );
    `

    `
    // Will only fire once when a unit is marked as complete, should not trigger again.
    do_action( 'coursepress_student_course_unit_completed', $student_id, $course_id, $unit_id );
    `