Reworked enrollment settings and signup button
Made enrollment settings ("Who can enroll?") more extensible and
added options for registered users. Solves concerns for allowing
anyone to register on sites.

Reworked the signup button. Simplified the number of options to
check and introduced $buttons array for adding the possible buttons
that can be rendered.  Also filtered for extensibility.

Fixed issue with pre-requisites that didn't work as expected.