Simplify/unify the logic to get custom post-types/taxonomy slugs plugin-wide
Before this change many classes had duplicate logic to add a custom prefix to
post-type slugs. To avoid compatibility issues all those functions now use the
same function to prefix the slug that is also used to register the post type.
Also we removed the param `$with_prefix` because it was never used.